+++ core_client/base/models/v3/process_config.py
@@ -29,6 +29,7 @@ class ProcessConfig(BaseModel):
     {
         "scheduler": str,
         "runtime_duration_seconds": int,
+        "log_pattern": [str]
     }
     """
 
@@ -44,6 +45,7 @@ class ProcessConfig(BaseModel):
     stale_timeout_seconds: int = 10
     scheduler: Optional[str]
     runtime_duration_seconds: Optional[int]
+    log_pattern: Optional[List[str]]
     type: ProcessConfigType = ProcessConfigType.ffmpeg
 
     class Config: