+++ core_client/base/models/v3/process_state_progress_io_framerate.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel
+
+
+class ProcessStateProgressIOFramerate(BaseModel):
+    """
+    {
+        "avg": 24.999,
+        "min": 25,
+        "max": 19.531,
+    }
+    """
+
+    avg: float
+    min: float
+    max: float